+++ core/build.sh
@@ -38,16 +38,22 @@ case $BUILDER_OS in
       ":x64    64-bit Windows (x64) build"
     )
     ;;
-  mac|linux)
+  mac)
     archtargets+=(
-      ":arch   Linux or mac build -- current architecture"
+      ":mac           Mac all architectures fat library build"
+      ":mac-x86_64    Mac Intel build"
+      ":mac-arm64     Mac arm64 (M1) build"
+    )
+    ;;
+  linux)
+    archtargets+=(
+      ":arch   Linux build -- current architecture"
     )
     ;;
 esac
 
-# TODO: consider using "linux" and "mac" instead of "arch"?
+# TODO: consider using "linux" instead of "arch"?
 #  ":linux          Build for current Linux architecture"
-#  ":mac            Build for current macOS architecture"
 
 builder_describe \
 "Build Keyman Core
@@ -63,11 +69,10 @@ Libraries will be built in 'build/<target>/<configuration>/src'.
   "configure" \
   "build" \
   "test" \
-  "install         install libraries to current system" \
-  "uninstall       uninstall libraries from current system" \
+  "install                         install libraries to current system" \
+  "uninstall                       uninstall libraries from current system" \
   "${archtargets[@]}" \
-  "--debug,-d                      configuration is 'debug', not 'release'" \
-  "--no-tests      do not configure tests (used by other projects)" \
+  "--no-tests                      do not configure tests (used by other projects)" \
   "--target-path=opt_target_path   override for build/ target path" \
   "--test=opt_tests,-t             test[s] to run (space separated)"
 
@@ -89,23 +94,38 @@ if builder_is_dep_build || builder_has_option --no-tests; then
   builder_remove_dep /developer/src/kmc
 fi
 
-if builder_has_option --debug; then
+if builder_is_debug_build; then
   CONFIGURATION=debug
 else
   CONFIGURATION=release
 fi
 
+# 'mac' target builds both x86_64 and arm architectures and
+# generates a 'fat' library from them.
+builder_describe_internal_dependency \
+  build:mac build:mac-x86_64 \
+  build:mac build:mac-arm64
+
 builder_describe_outputs \
-  configure:x86      build/x86/$CONFIGURATION/build.ninja \
-  configure:x64      build/x64/$CONFIGURATION/build.ninja \
-  configure:arch     build/arch/$CONFIGURATION/build.ninja \
-  configure:wasm     build/wasm/$CONFIGURATION/build.ninja \
-  build:x86          build/x86/$CONFIGURATION/src/libkmnkbp0.a \
-  build:x64          build/x64/$CONFIGURATION/src/libkmnkbp0.a \
-  build:arch         build/arch/$CONFIGURATION/src/libkmnkbp0.a \
-  build:wasm         build/wasm/$CONFIGURATION/src/libkmnkbp0.a
+  configure:x86             build/x86/$CONFIGURATION/build.ninja \
+  configure:x64             build/x64/$CONFIGURATION/build.ninja \
+  configure:mac             build/mac/$CONFIGURATION/ \
+  configure:mac-x86_64      build/mac-x86_64/$CONFIGURATION/build.ninja \
+  configure:mac-arm64       build/mac-arm64/$CONFIGURATION/build.ninja \
+  configure:arch            build/arch/$CONFIGURATION/build.ninja \
+  configure:wasm            build/wasm/$CONFIGURATION/build.ninja \
+  build:x86                 build/x86/$CONFIGURATION/src/libkmnkbp0.a \
+  build:x64                 build/x64/$CONFIGURATION/src/libkmnkbp0.a \
+  build:mac                 build/mac/$CONFIGURATION/libkmnkbp0.a \
+  build:mac-x86_64          build/mac-x86_64/$CONFIGURATION/src/libkmnkbp0.a \
+  build:mac-arm64           build/mac-arm64/$CONFIGURATION/src/libkmnkbp0.a \
+  build:arch                build/arch/$CONFIGURATION/src/libkmnkbp0.a \
+  build:wasm                build/wasm/$CONFIGURATION/src/libkmnkbp0.a
 
 # Target path is used by Linux build, e.g. --target-path keyboardprocessor
+# TODO: sort out builder_describe_outputs and TARGET_PATH -- preferably by
+#       having Keyman for Linux build Core normally and then _copy_ the
+#       required files into its target path, and eliminating --target-path
 if builder_has_option --target-path; then
   TARGET_PATH="$opt_target_path"
 else
@@ -115,7 +135,7 @@ fi
 # Iterate through all possible targets; note that targets that cannot be built
 # on the current platform have already been excluded through the archtargets
 # settings above
-targets=(wasm x86 x64 arch)
+targets=(wasm x86 x64 mac-x86_64 mac-arm64 arch)
 
 for target in "${targets[@]}"; do
   MESON_PATH="$TARGET_PATH/$target/$CONFIGURATION"
@@ -127,3 +147,28 @@ for target in "${targets[@]}"; do
   do_install $target
   do_uninstall $target
 done
+
+# After we have built the necessary internal dependencies, then we can go
+# ahead and build a fat library for external consumption
+if builder_start_action configure:mac; then
+  mkdir -p "$TARGET_PATH/mac/$CONFIGURATION"
+  builder_finish_action success configure:mac
+fi
+
+if builder_start_action build:mac; then
+  lipo -create \
+    "$TARGET_PATH/mac-x86_64/$CONFIGURATION/src/libkmnkbp0.a" \
+    "$TARGET_PATH/mac-arm64/$CONFIGURATION/src/libkmnkbp0.a" \
+    -output "$TARGET_PATH/mac/$CONFIGURATION/libkmnkbp0.a"
+  builder_finish_action success build:mac
+fi
+
+if builder_start_action test:mac; then
+  # We can only run the tests for the current architecture; we can
+  # assume that build:mac has run so both architectures will be
+  # available
+  target=mac-`uname -m`
+  MESON_PATH="$TARGET_PATH/$target/$CONFIGURATION"
+  meson test -C "$MESON_PATH" "${builder_extra_params[@]}"
+  builder_finish_action success test:mac
+fi